#72 refac: Widget 폴더명 및 폴더링 변경, MonthWidgetProvider 파일 분리
