+++ Widget/Assets.xcassets/AccentColor.colorset/Contents.json
