+++ Widget/Assets.xcassets/AppIcon.appiconset/Contents.json
