+++ Widget/Assets.xcassets/Background.colorset/Contents.json
