+++ Widget/Assets.xcassets/Background2.colorset/Contents.json
