+++ Widget/Assets.xcassets/Contents.json
