+++ Widget/Assets.xcassets/D10.colorset/Contents.json
