+++ Widget/Assets.xcassets/D11.colorset/Contents.json
