+++ Widget/Assets.xcassets/D2.colorset/Contents.json
