+++ Widget/Assets.xcassets/D3.colorset/Contents.json
