+++ Widget/Assets.xcassets/D5.colorset/Contents.json
