+++ Widget/Assets.xcassets/D6.colorset/Contents.json
