+++ Widget/Assets.xcassets/D7.colorset/Contents.json
