+++ Widget/Assets.xcassets/D9.colorset/Contents.json
