+++ Widget/Assets.xcassets/WidgetBackground.colorset/Contents.json
