+++ Widget/Assets.xcassets/tintColor.colorset/Contents.json
