+++ Widget/Assets.xcassets/titiIcon.imageset/Contents.json
