+++ Widget/Info.plist
