+++ Widget/Widgets/MonthWidget/MonthWidgetEntry.swift
@@ -9,6 +9,7 @@
 import Foundation
 import WidgetKit
 
+// MARK: MonthWidget 표시를 위한 데이터 부분
 struct MonthWidgetEntry: TimelineEntry {
     let date: Date
     let configuration: ConfigurationIntent
