+++ Widget/Widgets/MonthWidget/MonthWidgetProvider.swift
@@ -1,0 +1,37 @@
+//
+//  MonthWidgetProvider.swift
+//  Project_Timer
+//
+//  Created by Kang Minsang on 2023/05/11.
+//  Copyright © 2023 FDEE. All rights reserved.
+//
+
+import Foundation
+import WidgetKit
+
+// MARK: Timeline 및 Entry 반환 부분
+struct MonthWidgetProvider: IntentTimelineProvider {
+    func placeholder(in context: Context) -> MonthWidgetEntry {
+        MonthWidgetEntry(date: Date(), configuration: ConfigurationIntent())
+    }
+
+    func getSnapshot(for configuration: ConfigurationIntent, in context: Context, completion: @escaping (MonthWidgetEntry) -> ()) {
+        let entry = MonthWidgetEntry(date: Date(), configuration: configuration)
+        completion(entry)
+    }
+
+    func getTimeline(for configuration: ConfigurationIntent, in context: Context, completion: @escaping (Timeline<MonthWidgetEntry>) -> ()) {
+        var entries: [MonthWidgetEntry] = []
+
+        // Generate a timeline consisting of five entries an hour apart, starting from the current date.
+        let currentDate = Date()
+        for hourOffset in 0 ..< 5 {
+            let entryDate = Calendar.current.date(byAdding: .hour, value: hourOffset, to: currentDate)!
+            let entry = MonthWidgetEntry(date: entryDate, configuration: configuration)
+            entries.append(entry)
+        }
+
+        let timeline = Timeline(entries: entries, policy: .atEnd)
+        completion(timeline)
+    }
+}
