+++ Widget/ko.lproj/Localizable.strings
